\name{chronobiology-package}
\alias{chronobiology-package}
\docType{package}
\title{
\packageTitle{chronobiology}
}
\description{
\packageDescription{chronobiology}
}
\details{
Many species show diurnal movement pattern due to different lighting conditions. Additionall some species adapt to repeated disturbance events especially when time and place is predictable (e.g. hikers travelling along a path). Following this it could be of interest to classify your data according to specified points in time (e.g. Sunday noon, time before and after sunrise/sunset, night or day...). The functions within the chronobiology-package provide additional information for further analysis of this topic.
}
\author{
\packageAuthor{chronobiology}

Maintainer: \packageMaintainer{chronobiology}
}
\references{
The formula within the function twilight() was adapted from the following webpage:
http://lexikon.astronomie.info/zeitgleichung/

The function stimuli_distance() integrates the function distm() from the geosphere package.
}
\keyword{chronobiology-package}
\keyword{chronobiology}
